custom/requirements.txt
beautifulsoup4==4.14.3
lxml==6.0.2
neo4j==6.0.3
openpyxl==3.1.5
pandas==2.3.3
python-dotenv==1.2.1
requests==2.32.5
torch==2.9.1
transformers==4.57.3